+++ src/scalar_kernels.cpp
@@ -45,28 +45,22 @@ namespace poly_paint::detail
 
         const std::uint32_t alpha = source.a;
         const std::uint32_t inverse_alpha = maximum_channel - alpha;
-        const std::uint32_t red_source_term =
-            static_cast<std::uint32_t>(source.r) * alpha + 128;
-        const std::uint32_t green_source_term =
-            static_cast<std::uint32_t>(source.g) * alpha + 128;
-        const std::uint32_t blue_source_term =
-            static_cast<std::uint32_t>(source.b) * alpha + 128;
+        const std::uint32_t red_source_term = source.r * alpha + 128;
+        const std::uint32_t green_source_term = source.g * alpha + 128;
+        const std::uint32_t blue_source_term = source.b * alpha + 128;
         for (const MutableRgbaSpan rgba : rgba_spans)
         {
             for (std::size_t offset = 0; offset < rgba.size(); offset += rgba_channel_count)
             {
-                std::uint32_t red = red_source_term +
-                    static_cast<std::uint32_t>(rgba[offset]) * inverse_alpha;
+                std::uint32_t red = red_source_term + rgba[offset] * inverse_alpha;
                 red += red >> 8;
                 rgba[offset] = static_cast<std::uint8_t>(red >> 8);
 
-                std::uint32_t green = green_source_term +
-                    static_cast<std::uint32_t>(rgba[offset + 1]) * inverse_alpha;
+                std::uint32_t green = green_source_term + rgba[offset + 1] * inverse_alpha;
                 green += green >> 8;
                 rgba[offset + 1] = static_cast<std::uint8_t>(green >> 8);
 
-                std::uint32_t blue = blue_source_term +
-                    static_cast<std::uint32_t>(rgba[offset + 2]) * inverse_alpha;
+                std::uint32_t blue = blue_source_term + rgba[offset + 2] * inverse_alpha;
                 blue += blue >> 8;
                 rgba[offset + 2] = static_cast<std::uint8_t>(blue >> 8);
             }
@@ -88,9 +82,8 @@ namespace poly_paint::detail
         {
             for (std::size_t channel = 0; channel < rgb_channel_count; ++channel)
             {
-                const int difference = static_cast<int>(left[offset + channel]) -
-                    static_cast<int>(right[offset + channel]);
-                total += static_cast<std::uint64_t>(difference < 0 ? -difference : difference);
+                const int difference = left[offset + channel] - right[offset + channel];
+                total += difference < 0 ? -difference : difference;
             }
         }
         return total;
